Remove alert severity filtering
Recently the T has started using severity `1`, which is documented as
being purely "informational", for disruptions which have a significant
impact on rides. We were previously filtering on `severity >= 3` to try
to declutter the calendar, but it seems the values really don't mean
much now. This was perhaps inevitable given there are few instances
where the MBTA itself filters on severity values, so there is not much
incentive to keep their use consistent.

Fortunately the more recently-added `transient?` filter can pick up some
of the slack in keeping the calendar clean.