Improve cache memory usage
The `cache` library never actually removes expired entries (and in fact
does not even have a method to do so on demand), so it was possible for
the app to use unbounded memory. This attempts to address the issue by:

* Replacing `cache` with `memory_cache`, which has a method for cleaning
  up expired entries. This library can also use an arbitrary type as the
  key instead of requiring `String`, which allows simplifying some code.

* Introducing `TCal::Cache`, a refinement of `MemoryCache` which uses
  `tasker` to periodically clean up expired entries in the background.

`MemoryCache` also deletes expired entries on read, but periodic cleanup
is still needed when it's possible for an entry to be stored and then
never read again (for example, the site HTML cache, which is keyed with
the current date).